Update to enable file sharing.

+++ file_share/routes.py
@@ -1,8 +1,12 @@
 import os
+from itertools import zip_longest
+from datetime import datetime
+from .database import db_session
 from dotenv import load_dotenv
-from flask import render_template, redirect, request, Blueprint, send_file
+from flask import render_template, redirect, request, Blueprint, send_file, g, flash
 from .s3_functions import upload_file, download_file, list_files
 from flask_security import login_required
+from .models import Files
 
 load_dotenv()
 bucket = os.getenv('BACK_END_BUCKET')
@@ -10,12 +14,27 @@ bucket = os.getenv('BACK_END_BUCKET')
 bp = Blueprint('routes', __name__)
 
 
-@bp.route('/index')
+@bp.route('/')
+def home():
+    return redirect('index')
+
+
+@bp.route('/index', methods=['POST', 'GET'])
 @login_required
 def index():
-    """Home page rendering list of files with index.html file."""
-    files = list_files(bucket)
-    return render_template('index.html', contents=files)
+    """Index page rendering index.html file."""
+    all_files = list_files(bucket)
+    owned_files = []
+    shared_files = []
+    if all_files and g.user:
+        for file in all_files:
+            stored_files = Files.query.filter_by(name=file['Key']).all()
+            for stored_file in stored_files[-1:]:
+                if stored_file and stored_file.owner_id == g.user.id:
+                    owned_files.append(file)
+                elif stored_file and stored_file.shared and stored_file.owner_id != g.user.id:
+                    shared_files.append(file)
+    return render_template('index.html', owned_files=owned_files, shared_files=shared_files)
 
 
 @bp.route('/upload', methods=['POST'])
@@ -24,19 +43,49 @@ def upload():
     """Upload a file to the S3 bucket."""
     if request.method == 'POST':
         file = request.files['file']
-        file.save(file.filename)
-        upload_file(file.filename, bucket)
-        try:
-            os.remove(file.filename)
-        except OSError:
-            pass
+        if file.filename:
+            file.save(file.filename)
+            db_file = Files(name=file.filename,
+                            upload_time=str(datetime.now()),
+                            shared=False,
+                            owner_id=g.user.id)
+            db_session.add(db_file)
+            db_session.commit()
+            upload_file(file.filename, bucket)
+            try:
+                os.remove(file.filename)
+            except OSError:
+                pass
         return redirect('index')
 
 
-@bp.route('/download/<file_name>', methods=['GET'])
+@bp.route('/download/<file_name>', methods=['GET', 'POST'])
 @login_required
 def download(file_name):
     """Download a file from the S3 bucket to the local downloads folder."""
     if request.method == 'GET':
         output = download_file(file_name, bucket)
-        return send_file(filename_or_fp=output, as_attachment=True)
+        file = Files.query.filter_by(name=file_name).all()[-1]
+        file.last_download = str(datetime.now())
+        db_session.commit()
+        with open(output, 'w'):
+            return send_file(filename_or_fp=output, as_attachment=True), delete_file(file_name)
+
+
+def delete_file(file_name):
+    try:
+        os.remove(file_name)
+    except OSError:
+        pass
+
+
+@bp.route('/share', methods=['POST'])
+def share():
+    """Share files with other users."""
+    if request.method == 'POST':
+        shared_files = request.form.getlist('share_files')
+        for shared_file in shared_files:
+            file = Files.query.filter_by(name=shared_file).one()
+            file.shared = True
+            db_session.commit()
+        return redirect('index')
